+++ DumDum/DumDum/Models/JsonEntities/Troops/TroopsResponse.cs
@@ -8,8 +8,8 @@ namespace DumDum.Models.JsonEntities.Troops
         public string TroopType { get; set; }
         public int Level { get; set; }
         public int HP { get; set; }
-        public double Attack { get; set; }
-        public double Defence { get; set; }
+        public decimal Attack { get; set; }
+        public decimal Defence { get; set; }
         public int StartedAt { get; set; }
         public int FinishedAt { get; set; }
